+++ src/main/java/com/libraryquerypie/onlinelibrarysystem/entity/Borrow.java
@@ -1,5 +1,6 @@
 package com.libraryquerypie.onlinelibrarysystem.entity;
 
+import com.libraryquerypie.onlinelibrarysystem.enums.BorrowStatus;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
